Completed this HTML learnings project
I learned a lot in completing this HTML project, it gave me the feeling of being a true developer, looking up stuff on the internet to figure out stuff, researching how to center a div in HTML and so much more, furthermore I actually coded a whole bunch this time around without some Youtube tutorials by my side. This may have only been a pure HTML project, but it taught me a lot, and from now on I need to be more consistent in order to become a great developer in the future.